Add MySQL logging to SMTP server
Added functionality to log email data directly to a MySQL database in addition to writing the files to disk. This was achieved by introducing the 'mysql2' npm module and creating a connection to a local MySQL server. The code for the logging email data to MySQL database is wrapped in the onData callback function. Package dependencies in 'package.json' and 'package-lock.json' were updated accordingly with 'mysql2' module.